comit

+++ src/main/java/com/babas/utilitiesTables/buttonEditors/JButtonEditorPresentation.java
@@ -48,15 +48,17 @@ public class JButtonEditorPresentation extends AbstractCellEditor implements Tab
                 if (presentation.getProduct().getPresentations().size()>1){
                     boolean si=JOptionPane.showConfirmDialog(Utilities.getJFrame(),"¿Está seguro?, esta acción no se puede deshacer","Eliminar Presentación",JOptionPane.YES_NO_OPTION)==0;
                     if(si){
-                        presentation.refresh();
-                        presentation.getProduct().getPresentations().remove(presentation);
-                        if(presentation.isDefault()){
-                            presentation.getProduct().getPresentations().get(0).setDefault(true);
-                            presentation.getProduct().getPresentations().get(0).save();
-                            presentation.getProduct().setPresentationDefault(presentation.getProduct().getPresentations().get(0));
+                        if(presentation.getId()!=null){
+                            presentation.refresh();
+                            if(presentation.isDefault()){
+                                presentation.getProduct().getPresentations().get(0).setDefault(true);
+                                presentation.getProduct().getPresentations().get(0).save();
+                                presentation.getProduct().setPresentationDefault(presentation.getProduct().getPresentations().get(0));
+                            }
+                            presentation.getPrices().forEach(Babas::delete);
+                            presentation.delete();
                         }
-                        presentation.getPrices().forEach(Babas::delete);
-                        presentation.delete();
+                        presentation.getProduct().getPresentations().remove(presentation);
                         Notify.sendNotify(Utilities.getJFrame(), Notify.Type.SUCCESS, Notify.Location.TOP_CENTER,"ÉXITO","Presentación eliminada");
                     }
                 }else{
